refactor: 💡 move settings composables to feature folder
