+++ app/src/main/java/com/rickyhu/hushkeyboard/navigation/SettingsRoute.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.rickyhu.hushkeyboard.ui.settings.SettingsScreen
+import com.rickyhu.hushkeyboard.settings.SettingsScreen
 
 const val settingsRoute = "settings"
