+++ app/src/main/java/com/rickyhu/hushkeyboard/settings/ui/VibrateOnTapSwitchItem.kt
@@ -1,4 +1,4 @@
-package com.rickyhu.hushkeyboard.ui.settings.composables
+package com.rickyhu.hushkeyboard.settings.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.material3.Icon
@@ -13,17 +13,17 @@ import com.rickyhu.hushkeyboard.R
 import com.rickyhu.hushkeyboard.ui.theme.HushKeyboardTheme
 
 @Composable
-fun AddSpaceBetweenNotationSwitchItem(
+fun VibrateOnTapSwitchItem(
     value: Boolean,
     onValueChanged: (Boolean) -> Unit = {}
 ) {
     ListItem(
         modifier = Modifier.clickable { onValueChanged(!value) },
-        headlineContent = { Text("Add space after notation") },
+        headlineContent = { Text("Vibrate on tap") },
         leadingContent = {
             Icon(
-                painter = painterResource(R.drawable.ic_space),
-                contentDescription = "Space"
+                painter = painterResource(R.drawable.ic_vibration),
+                contentDescription = "Vibrate"
             )
         },
         trailingContent = {
@@ -34,9 +34,9 @@ fun AddSpaceBetweenNotationSwitchItem(
 
 @Preview(showBackground = true)
 @Composable
-fun AddSpaceBetweenNotationSwitchItemPreview() {
+fun VibrateOnTapSwitchItemPreview() {
     HushKeyboardTheme {
-        AddSpaceBetweenNotationSwitchItem(
+        VibrateOnTapSwitchItem(
             value = true
         )
     }
